# packages/protocol/openapi.yaml
openapi: 3.1.0
info:
  title: OpenOS osd API
  version: 0.1.0
  description: >
    Localhost-only system API served by the osd daemon. All endpoints require
    the per-session bearer token; the WebSocket authenticates via ?token=.
servers:
  - url: http://127.0.0.1:7700/api/v1
paths:
...
  /fs/list:
    get:
      summary: List a directory (confined to allowed roots)
      parameters: [{ name: path, in: query, schema: { type: string } }]
      responses: { '200': { description: '{path, entries: DirEntry[]}' }, '403': { description: outside allowed roots } }
  /fs/read:
    get:
      summary: Read a text file (8 MB cap)
      parameters: [{ name: path, in: query, required: true, schema: { type: string } }]
      responses: { '200': { description: text/plain body } }
...
  /office/status:
    get: { summary: Local CryptPad availability, responses: { '200': { description: '{url, running}' }, '404': { description: not configured } } }
...